rename: states 폴더 정리
