+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import { Suspense, lazy, useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
-import { isLoggedInState } from 'states/atoms';
+import { isLoggedInState } from 'states/auth';
 import { authAPI } from 'apis/authAPI';
 import Layout from 'routes/Layout';
 import Home from 'routes/Home';
