+++ client/src/components/common/PublicRoute/index.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from 'recoil';
-import { isLoggedInState } from 'states/atoms';
+import { isLoggedInState } from 'states/auth';
 import { Navigate, Outlet } from 'react-router-dom';
 
 const PublicRoute = () => {
